+++ source/app/service-providers/commands/export.ts
@@ -71,7 +71,22 @@ export default class Export extends ZettlrCommand {
       if (fileDescriptor !== null) {
         exporterOptions.sourceFiles.push(fileDescriptor)
         exporterOptions.cwd = fileDescriptor.dir
-        exporterOptions.targetDirectory = (exportTo === 'temp') ? app.getPath('temp') : fileDescriptor.dir
+        switch (exportTo) {
+          case 'ask': {
+            const folderSelection = await this._app.windows.askDir(trans('system.export_dialog.title'), null, trans('system.export_dialog.save'))
+            if (folderSelection === undefined || folderSelection.length === 0) {
+              this._app.log.error('[Export] Could not run exporter: Folderselection did not have a result!')
+              return
+            }
+            exporterOptions.targetDirectory = folderSelection[0]
+            break
+          }
+          case 'temp':
+          case 'cwd':
+          default:
+            exporterOptions.targetDirectory = (exportTo === 'temp') ? app.getPath('temp') : fileDescriptor.dir
+            break
+        }
       }
     }
 
@@ -101,7 +116,7 @@ export default class Export extends ZettlrCommand {
         }
       } else {
         const title = trans('system.error.export_error_title')
-        const message = trans('system.error.export_error_message', output.stderr[0])
+        const message = trans('system.error.export_error_message', `Pandoc exited with code ${output.code}`)
         const contents = output.stderr.join('\n')
         this._app.windows.showErrorMessage(title, message, contents)
       }